Add files via upload
A simple Quiz game project in JAVA

A simpple quiz game in java that has multiple choice questions and one correct answer for each with an option to bookmark the question and a score after the test is complete.

To run this code, 1.) Download the .java file and save it in a folder. 2.) This code will run in any IDE or even in JDK (Java Development Kit) via the command prompt. 3.) To run in IDE, simply copy paste the code then compile and then run. 4.) To run in JDK, copy paste the code in notepad then save the file as a .java file, open command prompt, go to the location where the file is(Eg: cd Desktop etc) type in 'javac Quiz.java' this will compile the file and check for errors, after this step type in java 'Quiz.java.' This will run the code.

Some JAVA IDE that can be used other than JDK, Blue J, Eclipse, InteliJ, IDEA Netbeans.